refactor: vendor form by splitting the section into separate files
